Brings the ~25-commit main advance (native test framework replacing GoogleTest,
unified command system, ban-store consolidation, map loading/Workshop, C++23,
doc rewrite, thread exception guards) together with this branch's netcode /
actor-replication / hardening work.

Resolution highlights (best-of-both, no fix dropped from either side):
- tests: ported all branch tests onto the native TestFramework.h drop-in macros;
  tests/CMakeLists.txt registers the union (26 sources), no gtest.
- BanManager: consolidated single-store (system_clock) + our Steam-id keying,
  duration clamp, stoi guard, don't-wipe-authoritative-list.
- ProtocolDecoder: main's sample-buffering decode architecture (already subsumes
  our defensive clamps).
- ConnectionManager: kept our ack-storm coalescing + two-packet pawn-spawn.
- ConfigManager: '#'/';' comments but never mid-value (security fix preserved).
- GameServer: command-system scaffold + our live attacker/victim weapon-fire guard
  (fixed an illegal 'continue' inside the new per-packet Guard lambda -> return).

Post-merge fixes (build was broken on MSVC, pre-existing on main):
- tests: /Zc:preprocessor so TestMock.h MOCK_METHOD detects 0-arg methods.
- MathUtilsTests: _USE_MATH_DEFINES before <cmath> for M_PI.
- CompressionHandlerTests: skip ZLIB cases when built without zlib.

Verified: full MSVC Debug build clean (0 errors), ctest 26/26 (100%) pass.

+++ src/Network/ConnectionManager.cpp
@@ -654,7 +654,10 @@ const std::vector<ActorBunchRecord>& GetActorBootstrapRecords() {
                               (static_cast<uint32_t>(all[off + 7]) << 8) |
                               (static_cast<uint32_t>(all[off + 8]) << 16) |
                               (static_cast<uint32_t>(all[off + 9]) << 24);
-            const size_t len = (r.bunchDataBits + 7) / 8;  // payload bytes
+            // Compute the payload byte count in size_t: doing (bunchDataBits + 7) in
+            // uint32 would wrap for bunchDataBits near UINT32_MAX, yielding a tiny len
+            // that passes the bounds check below and silently truncates the payload.
+            const size_t len = (static_cast<size_t>(r.bunchDataBits) + 7) / 8;  // payload bytes
             off += 10;
             if (off + len > all.size()) {
                 Logger::Warn("[ActorBootstrap] truncated record (bits=%u) - stopping", r.bunchDataBits);
